transport: add retry/backoff to stream send-loop; refactor
* common:
  - extend cos.IsRetriableConnErr to handle url.Error / net.Error
    with timeouts, plus ETIMEDOUT and ECONNABORTED
  - add cos.ClampDuration helper
* transport:
  - sendLoop change (major): add exponential backoff (with clamp and poor-man’s jitter)
  - stream collector: index streams by sessID instead of log string
  - switch log prefix from session ID to node ID for more meaningful log output
* misc ref:
  - rename constants for consistency (ini-blah-cap)
  - rename nlog.log => logf (as in: printf)

+++ cmn/nlog/api.go
@@ -20,15 +20,15 @@ const (
 var LogToStderr bool
 var MaxSize int64 = 4 * 1024 * 1024 // usually, config.log.max_size
 
-func InfoDepth(depth int, args ...any)    { log(sevInfo, depth, "", args...) }
-func Infoln(args ...any)                  { log(sevInfo, 0, "", args...) }
-func Infof(format string, args ...any)    { log(sevInfo, 0, format, args...) }
-func WarningDepth(depth int, args ...any) { log(sevWarn, depth, "", args...) }
-func Warningln(args ...any)               { log(sevWarn, 0, "", args...) }
-func Warningf(format string, args ...any) { log(sevWarn, 0, format, args...) }
-func ErrorDepth(depth int, args ...any)   { log(sevErr, depth, "", args...) }
-func Errorln(args ...any)                 { log(sevErr, 0, "", args...) }
-func Errorf(format string, args ...any)   { log(sevErr, 0, format, args...) }
+func InfoDepth(depth int, args ...any)    { logf(sevInfo, depth, "", args...) }
+func Infoln(args ...any)                  { logf(sevInfo, 0, "", args...) }
+func Infof(format string, args ...any)    { logf(sevInfo, 0, format, args...) }
+func WarningDepth(depth int, args ...any) { logf(sevWarn, depth, "", args...) }
+func Warningln(args ...any)               { logf(sevWarn, 0, "", args...) }
+func Warningf(format string, args ...any) { logf(sevWarn, 0, format, args...) }
+func ErrorDepth(depth int, args ...any)   { logf(sevErr, depth, "", args...) }
+func Errorln(args ...any)                 { logf(sevErr, 0, "", args...) }
+func Errorf(format string, args ...any)   { logf(sevErr, 0, format, args...) }
 
 func SetPre(dir, role string) {
 	logDir, aisrole = dir, role
